Move Annotation to graph.py
